compiler warnings
fixed over-aggressive casting sizeof() to float
a few other float to int and int to float

+++ freex3d/src/lib/scenegraph/Component_Networking.c
@@ -660,7 +660,7 @@ void load_Inline (struct X3D_Inline *node) {
 			} else if (res->status == ress_downloaded) {
 				res->media_type = resm_unknown;
 				res->whereToPlaceData = X3D_NODE(node);
-				res->offsetFromWhereToPlaceData = (float) offsetof (struct X3D_Inline, __children);
+				res->offsetFromWhereToPlaceData = offsetof (struct X3D_Inline, __children);
 				send_resource_to_parser(res);
 				node->__loadstatus = INLINE_PARSING; /* a "do-nothing" approach */
 			} else {
